requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
openpyxl
PyYAML
vaderSentiment
bertopic
sentence-transformers
umap-learn
hdbscan
gensim
